// backups/20260304_1507/types.ts
export interface Complaint {
  id: string;
  date: string;
  claimantName: string;
  claimantPaternal: string;
  claimantMaternal: string;
  claimantAddress: string;
  claimantDocType: string;
  claimantDocNumber: string;
  claimantPhone: string;
  claimantEmail: string;
  claimantAge: string;
  productService: 'PRODUCTO' | 'SERVICIO';
  claimedAmount: number;
  description: string;
  attendedBy?: string;
  type: 'RECLAMO' | 'QUEJA';
  detail: string;
  request: string;
  status: 'PENDIENTE' | 'ATENDIDO';
}